requirements.txt
google-genai
imageio==2.37.0
langchain_chroma==0.2.4
langchain_community==0.3.25
langchain_openai==0.3.24
matplotlib==3.10.3
numpy==1.23.5
openai==1.88.0
paddleocr==2.6.1.3
pandas==2.3.0
Pillow==11.2.1
pyautogui==0.9.54
python-dotenv==1.1.0
scipy==1.15.3
scikit-image
streamlit==1.46.0
supervision==0.25.1
tensorflow==2.19.0
termcolor==3.1.0
torch
torchvision
torchaudio
transformers==4.52.4
ultralytics==8.3.156
uvicorn==0.34.3
easyocr
paddlepaddle 